% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tstpr_commands.R
\name{tstp_assignattr}
\alias{tstp_assignattr}
\title{tstp_assignattr}
\usage{
tstp_assignattr(con, xts, tsid)
}
\arguments{
\item{con}{A tstp_connection object.}

\item{xts}{An xts object atttributes are assigned to.}

\item{tsid}{The timeseries ID.}
}
\value{
An xts object with assigned attributes.
}
\description{
Assigns tstp attributes to an xts-object
}
\details{
Assign tstp attributes to the xts-object provided.
}
\author{
Dominik Leutnant
}
\seealso{
\code{\link[xts]{xts}}
}

