% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tstpr_commands.R
\name{tstp_clearattr}
\alias{tstp_clearattr}
\title{tstp_clearattr}
\usage{
tstp_clearattr(xts)
}
\arguments{
\item{xts}{An xts object to be cleared.}
}
\value{
An xts object with empty tstp attributes.
}
\description{
Deletes attributes of a timeseries
}
\details{
Deletes tstp attibutes of the xts-object provided.
}
\seealso{
\code{\link[xts]{xts}}
}
\author{
Dominik Leutnant
}
