% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tstpr_class.R
\name{tstp_connection}
\alias{tstp_connection}
\title{tstp_connection}
\usage{
tstp_connection(host = NULL, port = NULL, user = "user", pass = "pass",
  group = NULL, verbose = FALSE, config.file = "~/.tstp.cnf",
  timeout = 5)
}
\arguments{
\item{host}{host}

\item{port}{port}

\item{user}{user}

\item{pass}{pass}

\item{group}{group}

\item{verbose}{logical. Provide additional details?}

\item{config.file}{The configuration file to be used if group is specified.}

\item{timeout}{Number of seconds to wait for a response until giving up. Can not be less than 1 ms.}
}
\description{
Creates a tstp_connection object
}
\seealso{
\code{\link[httr]{timeout}}
}
\author{
Dominik Leutnant
}
