% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tstpr_commands.R
\name{tstp_create}
\alias{tstp_create}
\title{tstp_create}
\usage{
tstp_create(con, xts)
}
\arguments{
\item{con}{A tstp_connection object.}

\item{xts}{An xts object with tstp attributes.}
}
\value{
A TSID (character) or 0.
}
\description{
Creates a timeseries
}
\details{
A complete attribute set encoded as URL is submitted. A TSID of the series 
which was created (or found in the database) is returned. In case the 
creation failed, the TSID obtains the value 0.
}
\author{
Dominik Leutnant
}
\seealso{
\code{\link[xts]{xts}}
}

