% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tstpr_commands.R
\name{tstp_delete}
\alias{tstp_delete}
\title{tstp_delete}
\usage{
tstp_delete(con, tsid)
}
\arguments{
\item{con}{A tstp_connection object.}

\item{tsid}{The timeseries ID.}
}
\value{
The server response is returned.
}
\description{
Deletes a timeseries
}
\details{
The TSID of the time series to be deleted is submitted.
}
\author{
Dominik Leutnant
}
