% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tstpr_commands.R
\name{tstp_deletequal}
\alias{tstp_deletequal}
\title{Deletes the quality level}
\usage{
tstp_deletequal(con, tsid, qual, from = NULL, to = NULL)
}
\arguments{
\item{con}{A tstp_connection object.}

\item{tsid}{The timeseries ID.}

\item{qual}{Sets the level of quality.}

\item{from}{Specifies the first date (\%Y-\%m-\%d).}

\item{to}{Specifies the last date (\%Y-\%m-\%d).}
}
\value{
The server response is returned.
}
\description{
A TSID, a time interval and a quality level are submitted. 
The quality is deleted in the interval range.
}
\author{
Dominik Leutnant
}

