% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tstpr_commands.R
\name{tstp_get}
\alias{tstp_get}
\title{tstp_get}
\usage{
tstp_get(con, tsid, from = NULL, to = NULL, qual = 1, type = "asc",
  verbose = TRUE, debug = FALSE, pm = FALSE)
}
\arguments{
\item{con}{A tstp_connection object.}

\item{tsid}{The timeseries ID.}

\item{from}{Specifies the first date (\%Y-\%m-\%d).}

\item{to}{Specifies the last date (\%Y-\%m-\%d).}

\item{qual}{Sets the level of quality}

\item{type}{Determines the way data are transmitted ('asc' or 'bin').}

\item{verbose}{logical. Provide additional details?}

\item{debug}{logical. For debugging purposes only.}

\item{pm}{logical. For performance meausurement only.}
}
\value{
A time series as an xts object.
}
\description{
Gets time series data from the tstp server
}
\details{
Gets time series data from the tstp server (datetime-value pairs).
Use GETCOMBO to get texts (datetime-string pairs). A time series is requested by
passing a TSID argument. In addition, a time interval with 
Von=<Point in time> and Bis=<Point in time> (from and to)  must be specified,
optionally a quality level with Qual=[0..47] can be specified. The highest 
quality level used is selected as default. A Base64 encoded binary block 
containing the list of the datetime-value pairs is delivered. 'Time reference',
'type of series' and 'unit' are written as XML attributes of the DEF element. 
Alternatively, the transmission can be carried out as ASCII which is 
selected via the parameter type=asc.
}
\author{
Dominik Leutnant
}
\seealso{
\code{\link[xts]{xts}}
}

