% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tstpr_commands.R
\name{tstp_getcombo}
\alias{tstp_getcombo}
\title{tstp_getcombo}
\usage{
tstp_getcombo(con, tsid, from = NULL, to = NULL, qual = 1,
  readmode = c("INTERPOLIERT", "INNEN", "AUSSEN"), debug = FALSE)
}
\arguments{
\item{con}{A tstp_connection object.}

\item{tsid}{The timeseries ID.}

\item{from}{Specifies the first date (\%Y-\%m-\%d).}

\item{to}{Specifies the last date (\%Y-\%m-\%d).}

\item{qual}{Sets the level of quality.}

\item{readmode}{Readmode.}

\item{debug}{logical. For debugging purposes only.}
}
\value{
A list of two xts objects.
}
\description{
Gets time series data from the tstp server (including texts)
}
\details{
The time series server stores three types of data lists: the list with numbers, the 
list with texts and the list with isolated points. GETCOMBO gets all three 
data lists in one single request. GETCOMBO behaves like GET.
The return comprises three consecutive TSD elements. The first is for numbers,
the second for texts, the third for isolated points. In addition to the 
parameters of the GET command (VON, BIS, QUAL) there is READMODE, 
that can take INTERPOLIERT (interpolated), INNEN (inside) or AUSSEN
(outside). The default is INTERPOLIERT.
}
\author{
Dominik Leutnant
}
\seealso{
\code{\link[xts]{xts}}
}

