% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tstpr_commands.R
\name{tstp_getdval}
\alias{tstp_getdval}
\title{tstp_getdval}
\usage{
tstp_getdval(con, tsid, ib, aus, from = NULL, to = NULL, verbose = FALSE,
  debug = FALSE)
}
\arguments{
\item{con}{A tstp_connection object.}

\item{tsid}{The timeseries ID.}

\item{ib}{Specifies the interval width (e.g., 5s, 30min, 1h, 7d, 1mon, 1a, ...)}

\item{aus}{Specifies the way values are consolidated (Sum, Mit, Min, Max, 
DMax, DMin, Lck)}

\item{from}{Specifies the first date (\%Y-\%m-\%d).}

\item{to}{Specifies the last date (\%Y-\%m-\%d).}

\item{verbose}{logical. Provide additional details?}

\item{debug}{logical. For debugging purposes only.}
}
\value{
A time series as an xts object.
}
\description{
Gets aggregated values of a specified interval
}
\details{
A TSID, a time interval with Von=<point in time> and Bis=<point in time>, 
an interval width with IB=<interval range> (e.g., 5s, 30min, 1h, 7d, 
1mon, 1a, ...), an interpretation with Aussage=<interpretation> (e.g. 
Mit(mean) - see chart 3 in appendix A) and optionally a quality layer with 
Qual=[0..50] are submitted. A time series with corresponding values
(interval values, continuous values or momentary values) is returned for the
time interval as binary block (see GET). The selected interval width IB 
defines the attributes time step factor and time step for the returned series.
}
\author{
Dominik Leutnant
}
