% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tstpr_commands.R
\name{tstp_getversion}
\alias{tstp_getversion}
\title{Gets versions available}
\usage{
tstp_getversion(con, xts)
}
\arguments{
\item{con}{A tstp_connection object.}

\item{xts}{An xts object with tstp attributes.}
}
\value{
A numeric vector with exisiting versions of the xts-object provided.
}
\description{
Returns a vector with exisiting versions of the xts-object provided
}
\author{
Dominik Leutnant
}
\seealso{
\code{\link[xts]{xts}}
}

