% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tstpr_commands.R
\name{tstp_glamp}
\alias{tstp_glamp}
\title{tstp_glamp}
\usage{
tstp_glamp(con, tsid, ib, from = NULL, to = NULL, verbose = FALSE)
}
\arguments{
\item{con}{A tstp_connection object.}

\item{tsid}{The timeseries ID.}

\item{ib}{Specifies the interval width (e.g., 5s, 30min, 1h, 7d, 1mon, 1a, 
...)}

\item{from}{Specifies the first date (\%Y-\%m-\%d).}

\item{to}{Specifies the last date (\%Y-\%m-\%d).}

\item{verbose}{logical. Provide additional details?}
}
\value{
A time series as an xts object.
}
\description{
Gets floating amplitudes
}
\details{
Calculates floating amplitutes (difference between the floating maxima and 
the floating minima). Supply a time series id (tsid), a focus with Von=time 
and Bis=time, and the width of the interval that is used to calculate the 
floating values (e.g. 1h): ib=width (e.g., 5s, 30min, 1h, 7d, 1mon, 1a, ...)
}
\author{
Dominik Leutnant
}
