% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tstpr_commands.R
\name{tstp_inspect}
\alias{tstp_inspect}
\title{tstp_inspect}
\usage{
tstp_inspect(con, tsid, from = NULL, to = NULL)
}
\arguments{
\item{con}{A tstp_connection object.}

\item{tsid}{The timeseries ID.}

\item{from}{Specifies the first date (\%Y-\%m-\%d).}

\item{to}{Specifies the last date (\%Y-\%m-\%d).}
}
\value{
The server response is returned.
}
\description{
Inspects a timeseries
}
\details{
Supply a time series id (tsid) and optionally a focus. Returned is the
maximal quality and the maximal physical quality on focus, the vita 
(Lebenslauf), the INFO field of the time series and the time of the last 
modification. Without a focus the maximal quality and maximal physical 
quality refer to the maxfocus of the time series.
}
\author{
Dominik Leutnant
}
