% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tstpr_commands.R
\name{tstp_put}
\alias{tstp_put}
\title{tstp_put}
\usage{
tstp_put(con, xts, text = FALSE)
}
\arguments{
\item{con}{A tstp_connection object.}

\item{xts}{An xts object to be uploaded.}

\item{text}{logical. Does the xts object contain characters?}
}
\value{
The server response is returned.
}
\description{
Puts an xts-object to the tstp server
}
\details{
A time series is selected by means of a TSID. Optionally, a quality 
level can be specified. The data is then submitted to the tstp server by POST
in an XML message and integrated on the server.
}
\author{
Dominik Leutnant
}

