% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tstpr_commands.R
\name{tstp_qnum}
\alias{tstp_qnum}
\title{tstp_qnum}
\usage{
tstp_qnum(con, tsid, from = NULL, to = NULL)
}
\arguments{
\item{con}{A tstp_connection object.}

\item{tsid}{The timeseries ID.}

\item{from}{Specifies the first date (\%Y-\%m-\%d).}

\item{to}{Specifies the last date (\%Y-\%m-\%d).}
}
\value{
The number of values of the focused time series is returned.
}
\description{
Queries the number of values
}
\details{
Supply a time series id (tsid) and optionally a focus. 
Returned is the number of quants (values) of the time series on focus. 
Without focus the total number is returned.
}
\author{
Dominik Leutnant
}

