% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tstpr_commands.R
\name{tstp_query}
\alias{tstp_query}
\title{tstp_query}
\usage{
tstp_query(con, ort = NULL, subort = NULL, param = NULL, defart = NULL,
  aussage = NULL, xdistanz = NULL, xfaktor = NULL, herkunft = NULL,
  reihenart = NULL, version = NULL, quelle = NULL, hauptreihe = NULL,
  zweck = NULL, ZRID = NULL, verbose = FALSE, verbose.cache = FALSE,
  debug = FALSE, pm = FALSE, all.attributes = FALSE)
}
\arguments{
\item{con}{A tstp_connection object.}

\item{ort}{query parameter (optional).}

\item{subort}{query parameter (optional).}

\item{param}{query parameter (optional).}

\item{defart}{query parameter (optional).}

\item{aussage}{query parameter (optional).}

\item{xdistanz}{query parameter (optional).}

\item{xfaktor}{query parameter (optional).}

\item{herkunft}{query parameter (optional).}

\item{reihenart}{query parameter (optional).}

\item{version}{query parameter (optional).}

\item{quelle}{query parameter (optional).}

\item{hauptreihe}{query parameter (optional).}

\item{zweck}{query parameter (optional).}

\item{ZRID}{query parameter (optional).}

\item{verbose}{logical. Provide additional details?}

\item{verbose.cache}{logical. Provide additional details of tstp cache?}

\item{debug}{logical. For debugging purposes only.}

\item{pm}{logical. For performance meausurement only.}

\item{all.attributes}{logical. Should all attributes be returned?}
}
\value{
A data.frame containing query results.
}
\description{
Queries the tstp server
}
\details{
An attribute pattern is sent to the tstp server. All time series matching the
pattern are returned to the client as a time series list together with all 
attributes and a number identifying it clearly (TSID) in XML format.
}
\author{
Dominik Leutnant
}

