% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tstpr_commands.R
\name{tstp_setattr}
\alias{tstp_setattr}
\title{tstp_setattr}
\usage{
tstp_setattr(con, tsid, attribute, value, verbose = FALSE)
}
\arguments{
\item{con}{A tstp_connection object.}

\item{tsid}{The timeseries ID.}

\item{attribute}{The attribute to be set.}

\item{value}{The attribute's value.}

\item{verbose}{logical. Should informative outputs printed during function evaluation?}
}
\value{
The server response is returned.
}
\description{
Sets attributes of a timeseries
}
\details{
An attribute of the time series selected by TSID is set. Identification 
attributes can not be set. Besides attributes you may pass INFO und 
LEBENSLAUF to set this fields of the time series.
}
\seealso{
\code{\link[xts]{xts}}
}
\author{
Dominik Leutnant
}
