% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tstpr_commands.R
\name{tstp_setvalue}
\alias{tstp_setvalue}
\title{Sets values to specified interval}
\usage{
tstp_setvalue(con, tsid, defart = c("K", "I", "M"), reihenart = c("Z", "R"),
  value = NA, from = NULL, to = NULL, by = "mins", tz = "GMT")
}
\arguments{
\item{con}{A tstp_connection object.}

\item{tsid}{The timeseries ID.}

\item{defart}{Specifies the defart.}

\item{reihenart}{Specifies the reihenart.}

\item{value}{The value to be set.}

\item{from}{Specifies the first date (\%Y-\%m-\%d).}

\item{to}{Specifies the last date (\%Y-\%m-\%d).}

\item{by}{Increment of the time sequence (e.g. "min")}

\item{tz}{Specifies the time zone.}
}
\value{
\code{\link[tstpr]{tstp_put}}
}
\description{
Sets values on a given time series within given range.
}
\note{
WARNING! This function overrides values without prompting!
}
\author{
Dominik Leutnant
}
