% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tstpr_class.R
\name{tstp_srv-class}
\alias{tstp_srv-class}
\title{tstp_srv class}
\description{
The tstp_srv class is used to communicate with the tstp server. 
It stores the tstp server connection information.
}
\section{Slots}{
 
 \describe{
   \item{\code{host}:}{Object of class \code{"character"} }
   \item{\code{port}:}{Object of class \code{"numeric"} }
   \item{\code{user}:}{Object of class \code{"character"} }
   \item{\code{pass}:}{Object of class \code{"character"} }
 }
}

\author{
Dominik Leutnant
}
