% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tstpr_commands.R
\name{tstp_update}
\alias{tstp_update}
\title{tstp_update}
\usage{
tstp_update(con, tsid)
}
\arguments{
\item{con}{A tstp_connection object.}

\item{tsid}{The timeseries ID.}
}
\value{
The server response is returned.
}
\description{
Forces the server to update a timeseries
}
\details{
A TSID is submitted. The server reads in the time series once again and 
updates the MaxFocus. In case the time series has been modified without 
the server having noticed this, this command must be applied.
}
\author{
Dominik Leutnant
}
