% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tstpr_upload.R
\name{upload_files}
\alias{upload_files}
\title{upload_files}
\usage{
upload_files(con, type, ort, subort, version, folder = NULL,
  use.pattern = TRUE, pattern = NULL, use.pattern.ex = FALSE,
  pattern.ex = NULL, upload = FALSE, parameter = NULL, max.files = 10)
}
\arguments{
\item{con}{A tstp.connection object.}

\item{type}{"wtw", "nivus" or "ott"}

\item{ort}{site}

\item{subort}{subsite}

\item{version}{version}

\item{folder}{folder}

\item{use.pattern}{logical. use.pattern}

\item{pattern}{pattern}

\item{use.pattern.ex}{logical. use.pattern.ex}

\item{pattern.ex}{pattern.ex}

\item{upload}{logical. Upload?}

\item{parameter}{Specifies the parameters to be read (currently all but 
nivus ignore this parameter).}

\item{max.files}{maximum number of files to upload.}
}
\value{
The names of uploaded xts-objects are invsibly returned.
}
\description{
Upload files: currently nivus, ott and wtw filetypes are supported
}
\details{
upload files to tstp
}
\author{
Dominik Leutnant
}
\keyword{internal}
