% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tstpr_upload.R
\name{upload_xts}
\alias{upload_xts}
\title{upload_xts}
\usage{
upload_xts(con, xts, text = FALSE)
}
\arguments{
\item{con}{A tstp.connection object.}

\item{xts}{An xts object to be uploaded.}

\item{text}{logical. Does the xts object contain characters?}
}
\value{
The server response is returned.
}
\description{
Uploads xts-objects
}
\details{
Upload of an xts-object
}
\author{
Dominik Leutnant
}
